Dat hang

+++ Backend/shopQuanAo/src/main/java/com/Nhom19/shopQuanAo/DTO/Response/Customer/MyOrder/OrderResponseDTO.java
@@ -1,11 +1,12 @@
 package com.Nhom19.shopQuanAo.DTO.Response.Customer.MyOrder;
 
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderResponseDTO(
         Integer maDdh,
         String tenKhachHang,
-        Double tongTien,
+        BigDecimal tongTien,
         List<OrderProductDTO> sanPham
 ) {}
